refactor(api): use directly test-utils/file.js in tests
